pyproject.toml
[project]
name = "pdf-to-vector-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colpali-engine>=0.3.6",
    "fastapi>=0.115.7",
    "fitz>=0.0.1.dev2",
    "pdf2image>=1.17.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.13.2",
    "sentence-transformers>=3.4.0",
    "spacy>=3.8.4",
    "stamina>=24.3.0",
    "torch>=2.5.1",
    "uvicorn>=0.34.0",
]
